+++ JobMatcherApp/urls.py
@@ -22,6 +22,4 @@ urlpatterns = [
     # view sets
     path('', include(router.urls)),
 
-    # Billing
-    path('charge/' , views.charge, name='charge')
 ]
