+++ billing/urls.py
@@ -1,0 +1,8 @@
+from django.urls import path
+
+from . import views
+
+urlpatterns = [
+    # Billing
+    path('charge/', views.charge, name='charge')
+]
